backend/requirements.txt
fastapi==0.123.0
uvicorn==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
PyPDF2==4.0.1
python-docx==0.8.11
